# By kk (9) and explon (8)
# Via explon
* develop:
  Lucene
  Lucene
  个人页面细节调整
  用户个人页面细节调整
  用户个人页面代码整理 细节调整
  用户个人页面代码优化
  Lucene
  个人页面修改提交错误合并
  Lucene
  Lucene
  增加更新文章搜索索引功能
  文章投票数据结构调整
  侧边栏可能感兴趣的人样式调整
  问题详情页侧边栏邀请bug修复
  搜索结果页面样式调整
  可能感兴趣的人或话题css调整
  限制后台只允许建立二级分类

+++ models/article.php
@@ -133,6 +133,11 @@ class article_class extends AWS_MODEL
 	
 	public function update_article($article_id, $title, $message, $topics, $create_topic)
 	{
+		if (!$article_info = $this->model('article')->get_article_info_by_id($article_id))
+		{
+			return false;
+		}
+		
 		$this->delete('topic_relation', 'item_id = ' . intval($article_id) . " AND `type` = 'article'");
 		
 		if (is_array($topics))
@@ -143,6 +148,14 @@ class article_class extends AWS_MODEL
 				
 				$this->model('topic')->save_topic_relation($this->user_id, $topic_id, $article_id, 'article');
 			}
+		}
+		
+		if (defined('G_LUCENE_SUPPORT') AND G_LUCENE_SUPPORT)
+		{
+			$this->model('search_lucene')->push_index('article', $title, intval($article_id), array(
+				'comments' => $article_info['comments'],
+				'views' => $article_info['views']
+			));
 		}
 		
 		return $this->update('article', array(
@@ -206,7 +219,7 @@ class article_class extends AWS_MODEL
 		), 'id = ' . intval($article_id));
 	}
 	
-	public function article_vote($type, $item_id, $rating, $uid, $reputation_factor)
+	public function article_vote($type, $item_id, $rating, $uid, $reputation_factor, $item_uid)
 	{
 		$this->delete('article_vote', "`type` = '" . $this->quote($type) . "' AND item_id = " . intval($item_id) . ' AND uid = ' . intval($uid));
 		
@@ -228,6 +241,7 @@ class article_class extends AWS_MODEL
 					'rating' => intval($rating),
 					'time' => time(),
 					'uid' => intval($uid),
+					'item_uid' => intval($item_uid),
 					'reputation_factor' => $reputation_factor
 				));
 			}
